Preprocess and dedup references before returning.
Preprocessing into groups is now supported in the data format (but not
yet done). Dedup is done, but should be moved after the sorting phase,
because now a deduped entry can show up before its original was seen,
which is not useful.

Apart from the deduping, will use grouping to group together files with
the same indicated match content. So the UI can avoid displaying
repetitive lines of same content.